Use new tooltip component
