+++ modules/features/referrals/src/main/kotlin/au/com/shiftyjelly/pocketcasts/referrals/ReferralsIconWithTooltip.kt
@@ -1,10 +1,7 @@
 package au.com.shiftyjelly.pocketcasts.referrals
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.padding
 import androidx.compose.material.Icon
 import androidx.compose.material.IconButton
 import androidx.compose.material.MaterialTheme
@@ -14,23 +11,17 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
+import androidx.compose.ui.unit.DpOffset
 import androidx.compose.ui.unit.dp
 import au.com.shiftyjelly.pocketcasts.compose.AppTheme
-import au.com.shiftyjelly.pocketcasts.compose.AppThemeWithBackground
 import au.com.shiftyjelly.pocketcasts.compose.CallOnce
 import au.com.shiftyjelly.pocketcasts.compose.Devices
 import au.com.shiftyjelly.pocketcasts.compose.LocalColors
 import au.com.shiftyjelly.pocketcasts.compose.ThemeColors
-import au.com.shiftyjelly.pocketcasts.compose.components.PopupDropdownMenuTooltip
-import au.com.shiftyjelly.pocketcasts.compose.components.TextH40
+import au.com.shiftyjelly.pocketcasts.compose.components.TipPosition
+import au.com.shiftyjelly.pocketcasts.compose.components.TooltipPopup
 import au.com.shiftyjelly.pocketcasts.compose.preview.ThemePreviewParameterProvider
 import au.com.shiftyjelly.pocketcasts.compose.theme
-import au.com.shiftyjelly.pocketcasts.payment.BillingCycle
-import au.com.shiftyjelly.pocketcasts.payment.SubscriptionOffer
-import au.com.shiftyjelly.pocketcasts.payment.SubscriptionPlans
-import au.com.shiftyjelly.pocketcasts.payment.SubscriptionTier
-import au.com.shiftyjelly.pocketcasts.payment.flatMap
-import au.com.shiftyjelly.pocketcasts.payment.getOrNull
 import au.com.shiftyjelly.pocketcasts.referrals.ReferralsViewModel.UiState
 import au.com.shiftyjelly.pocketcasts.ui.theme.Theme
 import au.com.shiftyjelly.pocketcasts.localization.R as LR
@@ -71,34 +62,14 @@ private fun ReferralsIconWithTooltip(
                 colors = LocalColors.current.colors,
             )
 
-            PopupDropdownMenuTooltip(
+            TooltipPopup(
                 show = state.showTooltip,
-            ) {
-                TooltipContent(
-                    referralPlan = state.referralPlan,
-                    onClick = onTooltipClick,
-                )
-            }
+                title = stringResource(LR.string.referrals_tooltip_message, state.referralPlan.offerDurationText),
+                tipPosition = TipPosition.TopStart,
+                anchorOffset = DpOffset(0.dp, -4.dp),
+                onClick = onTooltipClick,
+            )
         }
-    }
-}
-
-@Composable
-private fun TooltipContent(
-    referralPlan: ReferralSubscriptionPlan,
-    onClick: () -> Unit,
-) {
-    Column(
-        modifier = Modifier
-            .padding(horizontal = 16.dp)
-            .padding(top = 24.dp, bottom = 16.dp)
-            .clickable {
-                onClick()
-            },
-    ) {
-        TextH40(
-            text = stringResource(LR.string.referrals_tooltip_message, referralPlan.offerDurationText),
-        )
     }
 }
 
@@ -130,21 +101,5 @@ private fun IconWithBadgePreview(
                 colors = LocalColors.current.colors,
             )
         }
-    }
-}
-
-@Preview(device = Devices.PortraitRegular)
-@Composable
-fun TooltipContentPreview(
-    @PreviewParameter(ThemePreviewParameterProvider::class) themeType: Theme.ThemeType,
-) {
-    AppThemeWithBackground(themeType) {
-        TooltipContent(
-            referralPlan = SubscriptionPlans.Preview
-                .findOfferPlan(SubscriptionTier.Plus, BillingCycle.Yearly, SubscriptionOffer.Referral)
-                .flatMap(ReferralSubscriptionPlan::create)
-                .getOrNull()!!,
-            onClick = {},
-        )
     }
 }
